Remove convert_fontdb
